move CalculateHTTPStatusCode to errors package
